chore(inputs.cisco_telemetry_mdt): Factor out parsing to own file
